# runtime/swagger/http.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "MineAdmin API DOC",
        "description": "MineAdmin后台接口列表",
        "version": "1.1"
    },
    "servers": [
        {
            "url": "http://127.0.0.1:9501",
            "description": "本地服务器"
        }
    ],
    "paths": {
...
        "/system/app/index": {
            "get": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppIndexGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
        "/system/app/recycle": {
            "get": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRecycleGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/save": {
            "post": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppSavePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/read/{id}": {
            "get": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRead{id}GET",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/update/{id}": {
            "put": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppUpdate{id}PUT",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
        "/system/app/delete": {
            "delete": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppDeleteDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/realDelete": {
            "delete": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRealDeleteDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/recovery": {
            "put": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRecoveryPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/changeStatus": {
            "put": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppChangeStatusPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/app/remote": {
            "post": {
                "tags": [
                    "SystemAppController"
                ],
                "summary": "",
                "operationId": "SystemAppRemotePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
        "/system/user/index": {
            "get": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserIndexGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/recycle": {
            "get": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserRecycleGET",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/save": {
            "post": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserSavePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/read/{id}": {
            "get": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserRead{id}GET",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/update/{id}": {
            "put": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserUpdate{id}PUT",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/delete": {
            "delete": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserDeleteDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/realDelete": {
            "delete": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserRealDeleteDELETE",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/recovery": {
            "put": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserRecoveryPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/changeStatus": {
            "put": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserChangeStatusPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/clearCache": {
            "post": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserClearCachePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/setHomePage": {
            "post": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserSetHomePagePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/initUserPassword": {
            "put": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserInitUserPasswordPUT",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/updateInfo": {
            "post": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserUpdateInfoPOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/modifyPassword": {
            "post": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserModifyPasswordPOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/export": {
            "post": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserExportPOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/import": {
            "post": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserImportPOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/downloadTemplate": {
            "post": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserDownloadTemplatePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
        "/system/user/remote": {
            "post": {
                "tags": [
                    "UserController"
                ],
                "summary": "",
                "operationId": "SystemUserRemotePOST",
                "responses": {
                    "200": {
                        "description": "successful operation"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "500": {
                        "description": "System error"
                    }
                }
            }
        },
...
    },
    "components": {
...
        "securitySchemes": {
            "Authorization": {
                "type": "apiKey",
                "name": "Authorization",
                "in": "header"
            }
        }
    },
    "security": [
        {
            "Authorization": []
        }
    ],
...
}